pyproject.toml
[project]
name = "orchestra"
version = "0.1.0"
description = "MCP server testing and orchestration tool"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Ahaan Chaudhuri" }
]
keywords = ["mcp", "testing", "model-context-protocol", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "PyYAML>=6.0",
    "aiohttp>=3.9.0",
    "jsonpath-ng>=1.6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
orchestra = "orchestra.cli:app"

[project.urls]
Homepage = "https://github.com/ahaanc/orchestra"
Documentation = "https://github.com/ahaanc/orchestra#readme"
Repository = "https://github.com/ahaanc/orchestra"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestra"]

